+++ examples/resources/updown_recipient/resource.tf
@@ -1,0 +1,4 @@
+resource "updown_recipient" "myrecipient" {
+  type = "email"
+  value = "foo@bar.baz"
+}