+++ internal/provider/provider.go
@@ -25,9 +25,9 @@ func New() func() *schema.Provider {
 			},
 
 			ResourcesMap: map[string]*schema.Resource{
-				"updown_check":   checkResource(),
-				"updown_webhook": webhookResource(),
-                                "updown_recipients": recipientsResource()
+				"updown_check":     checkResource(),
+				"updown_webhook":   webhookResource(),
+				"updown_recipient": recipientResource(),
 			},
 		}
 	}